Refactor database config and update docs
